Created Customer Manager class to handle customers well implement serialization next
